+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -424,7 +424,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		}
 
 		@Override
-		public V get() {
+		protected V getValue() {
 			if (selectionEmpty.get()) {
 				return null;
 			}