+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityModelTest.java
@@ -192,7 +192,7 @@ public final class SwingEntityModelTest
 		}
 
 		@Override
-		public Entity get() {
+		protected Entity getValue() {
 			return comboBoxModel.selectedValue();
 		}
 	}