+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -618,7 +618,7 @@ public final class EntitySearchField extends HintTextField {
 		}
 
 		@Override
-		public String get() {
+		protected String getValue() {
 			return searchField.getText();
 		}